initial commit
main.py reads the weather from the openweathermap API using my
appid. It extracts the temperature, humidity, and wind info.

The thermocouple.py creates a connection to my Arduino Uno,
uploaded with a program to read the temperature every second.
The module takes the average temperature over 5 seconds/readings.

main.py takes that reading and writes it to a csv file along
with the weather data.